renamed darkside-tests/src/chain_generics.rs
